### 2022-04-30 #4361 @editorialbot
[PRE REVIEW]: DAISIEmainland: an R package for simulating an island-mainland system for macroevolution on islands
**Submitting author:** @joshwlambert (<a href="http://orcid.org/0000-0001-5218-3046">Joshua Lambert</a>)
**Repository:** https://github.com/joshwlambert/DAISIEmainland
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v1.0.0
