### 2022-04-30 #4361 @danielskatz
👋 @joshwlambert - where is the JOSS paper (following the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography))? I don't see it in the repo in the main or joss_submission branches
