### 2022-04-30 #4361 @joshwlambert
@danielskatz Hi, on the joss_submission branch the submodule DAISIEmainland_ms contains the paper.md, paper.bib and figures. Hope this helps, if a restructure would ease the review please let me know.
