### 2022-04-30 #4361 @danielskatz
Ah, sorry, I see it now
I'm not sure we can handle papers in submodules.
Can you just put it directly in the repo?
(@arfon - any other thoughts?)
