+++ JOSS/paper.bib
@@ -1,0 +1,127 @@
+
+@article{valente_equilibrium_2015,
+	title = {Equilibrium and non‐equilibrium dynamics simultaneously operate in the {Galápagos} islands},
+	volume = {18},
+	issn = {1461-023X, 1461-0248},
+	doi = {10.1111/ele.12461},
+	language = {en},
+	number = {8},
+	urldate = {2021-07-22},
+	journal = {Ecology Letters},
+	author = {Valente, Luis and Phillimore, Albert B. and Etienne, Rampal S.},
+	editor = {Harmon, Luke},
+	year = {2015},
+	pages = {844--852},
+	file = {Full Text:C\:\\Users\\P287218\\Zotero\\storage\\GXGJD8ME\\Valente et al. - 2015 - Equilibrium and non‐equilibrium dynamics simultane.pdf:application/pdf},
+}
+
+@article{valente_simple_2020,
+	title = {A simple dynamic model explains the diversity of island birds worldwide},
+	volume = {579},
+	issn = {0028-0836, 1476-4687},
+	doi = {10.1038/s41586-020-2022-5},
+	language = {en},
+	number = {7797},
+	urldate = {2021-07-22},
+	journal = {Nature},
+	author = {Valente, Luis and Phillimore, Albert B. and Melo, Martim and Warren, Ben H. and Clegg, Sonya M. and Havenstein, Katja and Tiedemann, Ralph and Illera, Juan Carlos and Thébaud, Christophe and Aschenbach, Tina and Etienne, Rampal S.},
+	year = {2020},
+	pages = {92--96},
+	file = {Full Text:C\:\\Users\\P287218\\Zotero\\storage\\BNNFVFNN\\Valente et al. - 2020 - A simple dynamic model explains the diversity of i.pdf:application/pdf;Full Text:C\:\\Users\\P287218\\Zotero\\storage\\GY44ZQFW\\Valente et al. - 2020 - A simple dynamic model explains the diversity of i.pdf:application/pdf},
+}
+
+@article{gillespie_general_1976,
+	title = {A general method for numerically simulating the stochastic time evolution of coupled chemical reactions},
+	volume = {22},
+	issn = {00219991},
+	doi = {10.1016/0021-9991(76)90041-3},
+	language = {en},
+	number = {4},
+	urldate = {2021-07-22},
+	journal = {Journal of Computational Physics},
+	author = {Gillespie, Daniel T.},
+	year = {1976},
+	pages = {403--434},
+}
+
+@article{gillespie_stochastic_2007,
+	title = {Stochastic {Simulation} of {Chemical} {Kinetics}},
+	volume = {58},
+	issn = {0066-426X, 1545-1593},
+	url = {http://www.annualreviews.org/doi/10.1146/annurev.physchem.58.032806.104637},
+	doi = {10.1146/annurev.physchem.58.032806.104637},
+	language = {en},
+	number = {1},
+	urldate = {2021-08-27},
+	journal = {Annual Review of Physical Chemistry},
+	author = {Gillespie, Daniel T.},
+	month = may,
+	year = {2007},
+	pages = {35--55},
+	file = {Full Text:C\:\\Users\\P287218\\Zotero\\storage\\QVTP8WP5\\Gillespie - 2007 - Stochastic Simulation of Chemical Kinetics.pdf:application/pdf},
+}
+
+@article{gillespie_exact_1977,
+	title = {Exact stochastic simulation of coupled chemical reactions},
+	volume = {81},
+	issn = {0022-3654, 1541-5740},
+	url = {https://pubs.acs.org/doi/10.1021/j100540a008},
+	doi = {10.1021/j100540a008},
+	language = {en},
+	number = {25},
[…]
