+++ JOSS/paper.md
@@ -1,0 +1,70 @@
+---
+title: 'DAISIEmainland: an R package for simulating an island-mainland system for macroevolution on islands'
+tags:
+  - R
+  - island biogeography
+  - macroevolution
+  - simulation
+authors:
+  - name: Joshua W. Lambert
+    orcid: 0000-0003-0872-7098
+    affiliation: 1
+  - name: Pedro Santos Neves
+    orcid: 0000-0003-2561-4677
+    affiliations: 1
+  - name: Rampal S. Etienne
+    orcid: 0000-0003-2142-7612
+    affiliations: 1
+  - name: Luis Valente
+    orcid: 0000-0003-4247-8785
+    affiliations: 1,2
+  - name: Richèl J.C. Bilderbeek
+    orcid: 0000-0003-0872-7098
+    affiliations: 1
+affiliations:
+ - name: Groningen Institute for Evolutionary Life Sciences, University of Groningen, Box 11103, 9700 CC Groningen, The Netherlands
+   index: 1
+ - name: Naturalis Biodiversity Center, Darwinweg 2, 2333 CR Leiden, The Netherlands
+   index: 2
+date: 15 December 2021
+bibliography: paper.bib
+---
+
+# Summary
+
+Speciation and extinction of species are two of the most fundamental processes that are investigated in the field of evolutionary biology. Ideally, one would like to study these processes in replicated isolated systems. Islands comes close to this setting if colonisation is rare. However, often we cannot directly measure when an island was colonised, when a colonist species speciated, or when species went extinct. We can, however, use genetic data to reconstruct the evolutionary history of an island community. With this reconstructed data we can fit statistical models to understand macroevolutionary processes on islands. These models make simplifying assumptions in order for the fitting procedure to be tractable. One such assumption, made by, for example, the DAISIE model (from the `DAISIE` R package [@etienne_daisie_2022]), is that mainland species (i.e. the pool of species that can colonise the island) cannot diversify or go extinct. `DAISIEmainland` is an R package that simulates speciation and extinction on the mainland as well as colonisation and diversification on the island. Providing a more realistic model of the island and the mainland for evolutionary biology research, `DAISIEmainland` features include: (1) simulate the evolutionary history on island species, (2) visualise that history, (3) calculate and plot summary metrics of the simulated data. The package enables the simulation of phylogenetic datasets from islands under a model more representative of biological systems to test current inference models in island biogeography (e.g. DAISIE).
+
+# Statement of Need
+
[…]
