+++ JOSS/README.md
@@ -1,0 +1,1 @@
+# DAISIEmainland article for Journal of Open Source Software (JOSS)
