+++ JOSS/joss_guidelines.md
@@ -1,0 +1,367 @@
+Submitting a paper to JOSS
+==========================
+
+If you've already developed a fully featured research code, released it under an [OSI-approved license](https://opensource.org/licenses), and written good documentation and tests, then we expect that it should take perhaps an hour or two to prepare and submit your paper to JOSS.
+But please read these instructions carefully for a streamlined submission.
+
+## Submission requirements
+
+- [x] The software must be open source as per the [OSI definition](https://opensource.org/osd).
+- [x] The software must have an **obvious** research application.
+
+- [x] You must be a major contributor to the software you are submitting, and have a GitHub account to participate in the review process.
+- [x] Your paper must not focus on new research results accomplished with the software.
+- [x] Your paper (`paper.md` and BibTeX files, plus any figures) must be hosted in a Git-based repository together with your software (although they may be in a short-lived branch which is never merged with the default).
+
+In addition, the software associated with your submission must:
+
+- [x] Be stored in a repository that can be cloned without registration.
+- [x] Be stored in a repository that is browsable online without registration.
+- [x] Have an issue tracker that is readable without registration.
+- [x] Permit individuals to create issues/file tickets against your repository.
+
+### What we mean by research software
+
+JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets; offers a mathematical library, or similar. While useful for many areas of research, pre-trained machine learning models are not in-scope for JOSS. 
+
+ * [x] RJCB: problem solved must be in text
+
+### Substantial scholarly effort
+
+ * [x] RJCB: yup, we do this
+
+JOSS publishes articles about software that represent substantial scholarly effort on the part of the authors. Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).
+
+As a rule of thumb, JOSS' minimum allowable contribution should represent **not less than** three months of work for an individual. Some factors that may be considered by editors and reviewers when judging effort include:
+
+- [x] Age of software (is this a well-established software project) / length of commit history.
+- [x] Number of commits.
+- [x] Number of authors.
+- [x] Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.
+- [x] Whether the software has already been cited in academic papers.
+
+> The DAISIEmainland package has been applied to test the robustness of the DAISIE model [@lambert_effect_2022] 
+
+- [x] Whether the software is sufficiently useful that it is _likely to be cited_ by your peer group.
+
+In addition, JOSS requires that software should be feature-complete (i.e., no half-baked solutions), packaged appropriately according to common community standards for the programming language being used (e.g., [Python](https://packaging.python.org), [R](https://r-pkgs.org/index.html)), and designed for maintainable extension (not one-off modifications of existing tools). "Minor utility" packages, including "thin" API clients, and single-function packages are not acceptable.
+
+### Co-publication of science, methods, and software
+
+ * [x] RJCB: mention the academic article
+
[…]
