### 2022-04-30 #4361 @joshwlambert
@danielskatz There is now a copy of the manuscript on the joss_submission branch in the JOSS dir (https://github.com/joshwlambert/DAISIEmainland/blob/joss_submission/JOSS/paper.md), the github actions will not trigger from this copy so if you would like to see whether the manuscript builds correctly please see https://github.com/joshwlambert/DAISIEmainland_ms/actions.
