### 2022-04-30 #4361 @danielskatz
👋 @vinisalazar - note that your paper does not compile.

The error that causes this is `Author (Pedro Santos Neves) is missing affiliation`

Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). In addition, you could work on the possibly missing DOI that editorialbot suggests, but note that this may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
