+++ JOSS/paper.bib
@@ -1,18 +1,60 @@
 
-@article{valente_equilibrium_2015,
-	title = {Equilibrium and non‐equilibrium dynamics simultaneously operate in the {Galápagos} islands},
-	volume = {18},
-	issn = {1461-023X, 1461-0248},
-	doi = {10.1111/ele.12461},
+@misc{etienne_daisie_2022,
+	title = {{DAISIE}: {Dynamical} {Assembly} of {Islands} by {Speciation}, {Immigration} and {Extinction}},
+	url = {https://doi.org/10.5281/zenodo.6474984},
+	author = {Etienne, Rampal S. and Valente, Luis M. and Phillimore, Albert B. and Haegeman, Bart and Lambert, Joshua W. and Neves, Pedro Santos and Xie, Shu and Bilderbeek, Richèl J. C. and Hildenbrandt, Hanno},
+	year = {2022},
+	annote = {R package version 4.1.1},
+}
+
+@misc{r_core_team_r_2022,
+	address = {Vienna, Austria},
+	title = {R: {A} {Language} and {Environment} for {Statistical} {Computing}},
+	url = {https://www.R-project.org/},
+	publisher = {R Foundation for Statistical Computing},
+	author = {{R Core Team}},
+	year = {2022},
+}
+
+@article{gillespie_general_1976,
+	title = {A general method for numerically simulating the stochastic time evolution of coupled chemical reactions},
+	volume = {22},
+	issn = {00219991},
+	doi = {10.1016/0021-9991(76)90041-3},
 	language = {en},
-	number = {8},
+	number = {4},
 	urldate = {2021-07-22},
-	journal = {Ecology Letters},
-	author = {Valente, Luis and Phillimore, Albert B. and Etienne, Rampal S.},
-	editor = {Harmon, Luke},
-	year = {2015},
-	pages = {844--852},
-	file = {Full Text:C\:\\Users\\P287218\\Zotero\\storage\\GXGJD8ME\\Valente et al. - 2015 - Equilibrium and non‐equilibrium dynamics simultane.pdf:application/pdf},
+	journal = {Journal of Computational Physics},
+	author = {Gillespie, Daniel T.},
+	year = {1976},
+	pages = {403--434},
+}
+
+@article{moran_random_1958,
+	title = {Random processes in genetics},
+	volume = {54},
+	issn = {0305-0041, 1469-8064},
+	doi = {10.1017/S0305004100033193},
+	language = {en},
+	number = {1},
+	urldate = {2021-10-19},
+	journal = {Mathematical Proceedings of the Cambridge Philosophical Society},
+	author = {Moran, P. A. P.},
+	month = jan,
+	year = {1958},
+	pages = {60--71},
+}
+
+@book{paradis_analysis_2006,
+	address = {New York},
+	series = {Use {R}!},
+	title = {Analysis of phylogenetics and evolution with {R}},
+	isbn = {978-0-387-32914-7},
+	publisher = {Springer},
+	author = {Paradis, Emmanuel},
+	year = {2006},
+	note = {OCLC: ocm71298831},
+	keywords = {Cladistic analysis, Methodology, Phylogeny, R (Computer program language)},
 }
 
 @article{valente_simple_2020,
@@ -28,20 +70,6 @@
 	year = {2020},
 	pages = {92--96},
 	file = {Full Text:C\:\\Users\\P287218\\Zotero\\storage\\BNNFVFNN\\Valente et al. - 2020 - A simple dynamic model explains the diversity of i.pdf:application/pdf;Full Text:C\:\\Users\\P287218\\Zotero\\storage\\GY44ZQFW\\Valente et al. - 2020 - A simple dynamic model explains the diversity of i.pdf:application/pdf},
-}
-
[…]
