### 2022-05-04 #4361 @joshwlambert
Hi @Bisaloo 😃 I do not have anyone personally in mind, but searching through the list of available JOSS reviewers it seems shaunpwilkinson, vbaliga and dombennett would be suitable reviewers given their programming language of choice and research topics.
