### 2022-05-05 #4361 @Bisaloo
Hi @shaunpwilkinson  :wave:, you signed up a while ago to review submissions to JOSS. Both the author and I think you would be a good fit for this submission. Would you be available to review this [R package](https://github.com/joshwlambert/DAISIEmainland) and [the associated paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04361/joss.04361/10.21105.joss.04361.pdf)?
