### 2022-05-05 #4361 @Bisaloo
Hi @TGuillerme :wave:, you are in our list of potential reviewers for JOSS. This submission seems in line with your interests. Would you be available to review this [R package](https://github.com/joshwlambert/DAISIEmainland) and [the associated paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04361/joss.04361/10.21105.joss.04361.pdf)?
