### 2022-05-09 #4361 @TGuillerme
Hi @Bisaloo (sorry for the delayed reply, I was on leave last week). Unfortunately I've already reviewed 4 manuscripts in the last two months so that's my quota done for a while. Maybe @alexsla (https://github.com/alexsla) would be a good fit (although I don't think he's in the list of reviewers for JOSS)?
