### 2022-05-26 #4361 @Bisaloo
Hi @ptitle, I found your contact while looking at BAMM and BAMMtools. Would you be interested in reviewing this [R package](https://github.com/joshwlambert/DAISIEmainland) and [the associated paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04361/joss.04361/10.21105.joss.04361.pdf) for JOSS? It seems very relevant to your interests.

> DAISIEmainland: an R package for simulating an island-mainland system for macroevolution on islands

You can have a look at our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected :relaxed:.
