### 2022-06-07 #4361 @Bisaloo
Hi @ohagen :wave:, would you be interested in reviewing this [R package](https://github.com/joshwlambert/DAISIEmainland) and [the associated paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04361/joss.04361/10.21105.joss.04361.pdf) for JOSS? Two persons (including the first author of the submission) recommended you for this submission and I do agree you would be a good fit.

> DAISIEmainland: an R package for simulating an island-mainland system for macroevolution on islands

You can have a look at our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected :relaxed:.
