### 2022-06-20 #4491 @Bisaloo
:wave: :wave: :wave: @joshwlambert, @alexskeels, @vboussange, this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
