### 2022-07-06 #4491 @vboussange
Hi everyone 👋
I am posting here my comments for the first round of review.

## General comments
### Reviewer summary
DAISIEmainland is a package that allows to simulate the evolutionary history of an island community, bringing more realism to the DAISIE model by simulating the evolutionary history of the mainland community.

### General comment
The package is well constructed, with a nice documentation, complete unit-tests, and community guidelines. However, I am not convinced of the usefulness of the package for the community, as it is presented right now. It seems that this package has been specifically designed to lead the investigation of the autor's work "The effect of mainland dynamics on data and parameter estimates in island biogeography", available as a preprint at https://www.biorxiv.org/content/10.1101/2022.01.13.476210v2. While I do not question the relevance of this study, I do not see the point of publishing the corresponding code as a standalone package. Even if the authors would less emphasize on the capacity of DAISIEmainland to test the DAISIE inference performance, and rather put emphasis on the capacity of DAISIEmainland to obtain more realistic evolutionary history simulations, I think that it would benefit the user to integrate those changes in the main DAISIE package, as most of the syntax is the same. Overall,  my opinion is that the code of DAISIEmainland could be a nice contribution to bring more functionalities to DAISIE, but it does not justify a standalone package.

## Specific comments
### Functionality
#### Installation
```
remotes::install_github("rsetienne/DAISIE")
```

Does unfortunately not work for me, I get an error with gfortran. The error is documented [here](https://www.cynkra.com/blog/2021-03-16-gfortran-macos/) but the fix did not work for me. It would be nice to indicate this potential fix to the user in the installation page.

### Documentation
#### Simulation algorithm
> Species that do not go extinct are **give** a spec_ex_t equal to the total time of the simulation.

> The algorithm checks whether any changes have occured on the mainland since the last time step and if so, the system is updated and the returned to the time at which the mainland last changed.

Not clear

> where r_j are the rates.

Breaking down the sentence leads to confusion, would be nice to rephrase this sentence.

> $X = \lambda ...$

I think that you are here confusing about the random variable which serves as time steps and the probability density. It would probably make more sense to denote this random variable $\Delta t$

#### Simulation data visualisation
> An island with a single endemic species **which whose** colonisation time is older than the island age so it is considered an island-age colonisation. The uncertainty in colonisation time is handled in an equivalent way to the non-endemic island-age colonisation.

> An island with **an recolonisation** of the same mainland species after it has colonised and speciated (either via cladogenesis or anagenesis on the island). The colonists that are from the same ancestral mainland species but the island species is now endemic **are show** by ‘Colonist after anagenesis’ and ‘Colonist after cladogenesis.’

#### Inference performance
> **simualation** algorithm

- Would be nice to have a link to the documentation of the DAISIE package, because it seems a prerequisite to understand the goal of the DAISIEmainland package.

- Why do you start the inference with initial parameter values which are exactly the same as those that were used to simulate the data?

- I would have liked to have here some comments about the performance assessed. What do those numbers tell? How can we interpret it?

### Paper
#### Summary
> 1) simulating the evolutionary history on island species,

Not sure what is meant by that.

#### Statement of need
> the performance and robustness of this island biogeography inference model is unknown when its assumptions are violated under biologically realistic scenarios
What is meant by "performance" here? "robustness"? I would simply delete this sentence.

> A central assumption of the DAISIE likelihood model

Not sure what is a likelihood model. Overall, l36-37 could be better formulated.

> The DAISIEmainland package has been applied to test the ... which provides a suite of phylogenetic likelihood inference models for island biogeography.

This should not belong to the statement of need. It rather describes the package, and I think that it would rather fit in the summary.

- Would be nice to explain in more details what's exactly meant by incomplete information data? What is exactly meant by ideal vs empirical data? Since I do not come from the empirical perspective, I find that the what is "empirical data" is not clear.
- Figure 2 caption is not clear: branch color represents the endemicity status, but in legend it says "endemic clade at maximum age".
