### 2022-07-29 #4491 @alexskeels
Hello @Bisaloo, @joshwlambert, @vboussange. I am very sorry for my lateness - i've been away on holidays.

A few technical issues prevented a complete review of the package. I ran into the folllowing error in vignette 4 running the ML inference:
"Error in DAISIE_ode_cs(initprobs, tvec, parsvec, atol, rtol, method, runmod = "daisie_runmod") :
  cpp_daisie_cs_runmod: too many steps"

and in vignettes 5 for summarising the results:
	DAISIEmainland::calc_num_spec doesn’t run “Error in stacs[[j]] : subscript out of bounds”
	DAISIEmainland::calc_num_col “Error in stacs[[j]] : subscript out of bounds”

I am running R version 4.1.2 on a windows machine if thats helpful at all. These should probably be addressed before assessing the checklist "Functionality" and "Performance".

One further general comment: The model simulates the mainland clade with an extinction rate and then speciation immediately follows extinction. How come the mainland pool has to remain a fixed size? Can it evolve under a BD model which might better represent flcutuating clade dyanmics on the mainland? I'm not asking for you to change anything of course, but I would be very interested to hear the justification (technical and biological) - perhaps a quick statement could be added to this end?
