### 2022-08-11 #4491 @joshwlambert
Hi @Bisaloo , @alexskeels, & @vboussange.

Firstly, thank you @vboussange for the comments. I will use the specific comments to improve the package.

@alexskeels Apologies for not replying sooner. I am finishing my PhD at the moment and will be submitting the next couple of weeks, and so have been extremely busy. I'm afraid I do not think I'll be able to get to the technical issues that prevented you from completing a full review for the next few weeks.

@Bisaloo Given the comments that have already been received and the fact that I will not be able to invest the time required to address the comments over the coming weeks, is it best to remove this from consideration for publication? I am happy to wait but do not want to waste the time of the reviewers and editors.
